InterAcT-GUI/requirement_file.txt
absl_py==1.4.0
joblib==1.2.0
matplotlib==3.7.1
numpy==1.20.3
optuna==3.2.0
pandas==2.0.1
PyYAML==6.0
scikit_learn==1.3.2
scipy==1.10.1
tensorflow==2.10.0
tensorflow_addons==0.20.0